face-attendance-pro/requirements.txt
fastapi
uvicorn
jinja2
pandas
pillow
face_recognition
opencv-python